requirements.txt
Package              Version
-------------------- -----------
adbutils             2.8.0
annotated-types      0.7.0
anyio                4.5.2
apkutils2            1.0.0
argon2-cffi          23.1.0
argon2-cffi-bindings 21.2.0
asgiref              3.5.2
bcrypt               4.2.1
beautifulsoup4       4.12.3
betterproto          1.2.5
black                24.8.0
blinker              1.4
Brotli               1.0.9
build                1.2.2.post1
CacheControl         0.14.2
certifi              2024.12.14
cffi                 1.17.1
chardet              5.2.0
charset-normalizer   3.4.1
cigam                0.0.3
cleo                 2.1.0
click                8.1.8
colorama             0.4.6
construct            2.10.70
crashtest            0.4.1
cryptography         36.0.2
decorator            5.1.1
Deprecated           1.2.15
deprecation          2.1.0
distlib              0.3.9
dnspython            2.6.1
dulwich              0.21.7
email_validator      2.2.0
et_xmlfile           2.0.0
exceptiongroup       1.2.2
fastapi              0.111.1
fastapi-cli          0.0.7
fastjsonschema       2.21.1
filelock             3.16.1
Flask                2.0.3
frida                16.5.9
frida-tools          13.6.0
grpclib              0.4.7
h11                  0.13.0
h2                   4.1.0
hpack                4.0.0
httpcore             1.0.7
httptools            0.6.4
httpx                0.28.1
hyperframe           6.0.1
idna                 3.10
imageio              2.35.1
importlib_metadata   8.5.0
importlib_resources  6.4.5
installer            0.7.0
itsdangerous         2.2.0
jaraco.classes       3.4.0
Jinja2               3.1.5
kaitaistruct         0.9
keyring              24.3.1
lazy_loader          0.4
ldap3                2.9.1
loguru               0.7.3
lxml                 5.3.0
markdown-it-py       3.0.0
MarkupSafe           2.1.5
mdurl                0.1.2
minio                7.2.15
mitmproxy            8.0.0
more-itertools       10.5.0
mpmath               1.3.0
msgpack              1.0.5
multidict            6.1.0
mypy-extensions      1.0.0
networkx             3.1
numpy                2.2.3
openpyxl             3.1.5
packaging            24.2
pandas               2.2.3
paramiko             3.5.1
passlib              1.7.4
pathspec             0.12.1
pexpect              4.9.0
pika                 1.3.2
pillow               10.4.0
pip                  25.0.1
pkginfo              1.12.0
platformdirs         4.3.6
poetry               1.8.5
poetry-core          1.9.1
poetry-plugin-export 1.8.0
prompt_toolkit       3.0.48
protobuf             3.19.0
psutil               7.0.0
ptyprocess           0.7.0
publicsuffix2        2.20191221
py                   1.11.0
pyasn1               0.6.1
pycparser            2.22
pycryptodome         3.21.0
pydantic             2.10.4
pydantic_core        2.27.2
pydivert             2.1.0
pyelftools           0.31
PyExecJS             1.5.1
Pygments             2.19.1
pymongo              4.10.1
PyNaCl               1.5.0
pyOpenSSL            22.0.0
pyparsing            3.0.9
pyperclip            1.8.2
pyproject_hooks      1.2.0
python-dateutil      2.9.0.post0
python-dotenv        1.0.1
python-multipart     0.0.20
pytz                 2025.1
PyWavelets           1.4.1
pywin32              308
pywin32-ctypes       0.2.3
PyYAML               6.0.2
rapidfuzz            3.9.7
requests             2.32.3
requests-toolbelt    1.0.0
retry                0.9.2
rich                 13.9.4
rich-toolkit         0.12.0
ruamel.yaml          0.17.40
ruamel.yaml.clib     0.2.8
scikit-image         0.21.0
scipy                1.10.1
scp                  0.15.0
setuptools           75.8.2
shellingham          1.5.4
six                  1.17.0
sniffio              1.3.1
sortedcontainers     2.4.0
soupsieve            2.6
starlette            0.37.2
stringcase           1.2.0
sympy                1.13.3
tenacity             9.0.0
tifffile             2023.7.10
tomli                2.2.1
tomlkit              0.13.2
tornado              6.4.2
tqdm                 4.67.1
trove-classifiers    2025.1.7.14
typer                0.15.1
typing_extensions    4.12.2
tzdata               2025.1
uiautodev            0.5.0
uiautomator          1.0.2
uiautomator2         3.2.8
urllib3              2.2.3
urwid                2.1.2
uvicorn              0.33.0
virtualenv           20.28.1
watchdog             4.0.2
watchfiles           0.24.0
wcwidth              0.2.13
wdapy                0.2.4
websockets           13.1
Werkzeug             2.2.2
wheel                0.45.1
win32_setctime       1.2.0
wrapt                1.17.0
wsproto              1.1.0
xmltodict            0.14.2
zipp                 3.20.2
zstandard            0.17.0
